move nested_with_bound into impl of PrettyCx
